Update language scripts to use the font
