--- Firmware/Fonts/FontTable.h
@@ -1,18 +1,0 @@
-{0x3E, {0x0F, 0xA8, 0xAF, 0x00, }, 'H'}, // '🄷', \x80
-{0x00, {0x69, 0x96, 0x00, 0x00, }, '\xdf'}, // '°', \x81
-{0x60, {0x25, 0x55, 0x58, 0x87, }, 'h'}, // '🌡', \x82
-{0x04, {0x27, 0xF2, 0xE0, 0x00, }, '^'}, // '⬏', \x83
-{0x34, {0x03, 0xCC, 0x19, 0x60, }, '\xf3'}, // '🔃', \x84
-{0x3C, {0x0E, 0xF8, 0x8F, 0x00, }, '\xdb'}, // '🗀', \x85
-{0x08, {0x02, 0x94, 0x92, 0x00, }, '>'}, // '»', \x86
-{0x1C, {0x07, 0x9A, 0x87, 0x00, }, '\xe5'}, // '🕑', \x87
-{0x24, {0x00, 0x85, 0x28, 0x52, }, '\x7e'}, // '⏬', \x88
-{0x06, {0x00, 0x1B, 0xE4, 0x00, }, '\x7e'}, // '✔', \x89
-{0xF1, {0x87, 0x07, 0x07, 0x87, }, 'a'}, // 'ă', \x8A
-{0xF0, {0x25, 0x07, 0x07, 0x87, }, 'a'}, // 'â', \x8B
-{0x00, {0x25, 0x06, 0x22, 0x27, }, 'i'}, // 'î', \x8C
-{0x10, {0x07, 0x87, 0x0F, 0x24, }, 's'}, // 'ș', \x8D
-{0x10, {0x4E, 0x44, 0x43, 0x24, }, 't'}, // 'ț', \x8E
-{0x00, {0x25, 0x07, 0x22, 0x27, }, 'I'}, // 'Î', \x8F
-{0x12, {0x78, 0x63, 0x87, 0x24, }, 'S'}, // 'Ș', \x90
-{0x01, {0xF2, 0x22, 0x20, 0x24, }, 'T'}, // 'Ț', \x91
